SSSCTL: Move sss_debuglevel to sssctl debug-level
Move code from sss_debuglevel to sssctl_logs.c and add new debug-logs
sssctl command to perform the same task of changing debug level
dynamically.

POPT_CONTEXT_KEEP_FIRST Flag added to poptGetContext call in
sssctl_debug_level() to fix argument parsing.

Resolves:
https://pagure.io/SSSD/sssd/issue/3057
